ppsthread.c: replace (long long) with (int64_t).
That makes the size just right for both 64-bit and 32-bit hosts.

+++ timespec.h
@@ -28,9 +28,9 @@
 /* return the difference between timespecs in nanoseconds
  * int may be too small, 32 bit long is too small, floats are too imprecise,
  * doubles are not quite precise enough 
- * MUST be long long to maintain precision on 32 bit code */
+ * MUST be at least int64_t to maintain precision on 32 bit code */
 #define timespec_diff_ns(x, y) \
-    (long long)((((x).tv_sec-(y).tv_sec)*NS_IN_SEC)+(x).tv_nsec-(y).tv_nsec)
+    (int64_t)((((x).tv_sec-(y).tv_sec)*NS_IN_SEC)+(x).tv_nsec-(y).tv_nsec)
 
 static inline void TS_NORM( struct timespec *ts)
 {
